Fall2022/programs/pa2/Zikmund_Noah/requirements.txt
attrs==22.1.0
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.3
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pyzmq==24.0.1
scipy==1.9.1
six==1.16.0
tomli==2.0.1
